tests/main/userns: ensure apparmor_parser doesn't pin the ABI
Instead specify to use the ABI presented by the kernel itself to ensure that the
parser does not silently downgrade the policy.

In the future I suspect we want snapd to always use the kernel ABI for
apparmor_parser and then store this ABI within the system-key to ensure policy
gets regenerated if / when the AppArmor kernel feature set changes.

Signed-off-by: Alex Murray <alex.murray@canonical.com>