prompting/storage: implemented decision pruning
When a new decision is added to the database via `Set()`, it should
overwrite previous decisions which are "more specific" than it.

That is, if the decision for access with directory is added for
`~/Documents`, then a previous decision about file access to
`~/Documents/foo.txt` should be deleted, regardless of what that prior
decision was, since the new decision applies to that path as well.

The guidlines for what makes a decision "more specific" are as follows:
- in all cases, delete all decisions which exactly match the path from
  all allow maps ("allow", "allow-directory", and "allow-subdirectories")
  - this is primarily because we cannot have duplicate paths appear in
    more than one allow map
- if the new decision is access to contents of a directory
  ("allow-directory"), then delete all decisions from the "allow" map
  for which the parent of the existing decision path is equal to the path
  of the new decision
- if the new decision is access to contents of a directory and all
  subdirectories ("allow-subdirectories"), then delete all decision from
  all maps for which the new decision path is an ancestor of the the
  existing decision path

Decision databases are completely independent between different
permissions (operations).  Thus, there may be "conflicting" decisions
for the same path if those decisions pertain to different permissions,
and "more specific" decisions from the databases corresponding to other
permissions are not pruned.

This change also adds deletion records to `Set()` in preparation for
future changes.  In particular, `Set()` now returns a `permissionDB` of
decision which have been deleted as a result of adding a new decision to
the database.  This may be useful for communicating to clients that
decisions have changed, or updating decision records that are indexed by
ID for multiple permissions at a time.  It will likely be desirable to
add an additional return field detailing the new decision added by
`Set()` (if any) for the same reason, particularly since `Set()` does
not add a new decision if it is already implied by a previous decision.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>