prompting/storage: added coalesce check for new decision
If the path for a new decision exactly matches that of a previous
decision (and that previous decision should be overwritten), then after
removing the old decision, it may be the case that the new decision is
implied by an older less-specific rule, and thus should not be added.

This change abstracts out the check for whether a decision is implied by
a previous decision so that it can be called in this context, as well as
the previous one, which is near the beginning of the permission loop
within `Set()`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>