UDENG-535: allow access to pipewire socket
This PR ports UDENG-418 to upstream, that enables audio in core
Desktop. It allows access to the pipewire socket from snapped
applications, and allows to have pipewire running inside a snap
container. To avoid security issues, it only allows to access
the sockets located at the folder belonging to the snap that
owns the slot.

With this patch, audio works fine in Core Desktop, and also
screen sharing.