github: fix/simplify `go fmt` check
It looks like the format check in the static tests got broken when
we moved to go 1.18. It was most likely missed because the code
does not check for `1.13` but uses a regexp `1\.13`.

This commit fixes this and also simplifies the code around this
so that on the next move to a newer go-version we have an easier
time finding the version.