cmd/snap-device-helper: handle bind and unbind commands
When a device is bound unound to/from a driver, BIND/UNBIND events are
sent. We have been generating error as "unknown action", which
pollutes the logs and confuse users.

While a device node should be present in theory already from the ADD
event, we still ensure the device is added to the cgroup on BIND
events. Like we do with CHANGE events.