interfaces/audio: remove write permissions in pulse folder (#12864)
Currently, any application that connects to the audio-playback or
the pulseaudio insterfaces, have write permissions to the pulse/
folder. This means that a snapped malicious application would be
able to delete the socket and put its own, being able to read the
audio of any other snapped application.

This patch removes those permissions.

* Remove unneeded rules

The "owner /run/pulse/native/ rwk" rule is already managed by a
similar previous one. Also, there is no need to allow to link the
pulse folder.

Tested both in Firefox and Telegram, and everything still works
fine.