tests: add test that explores `apparmor_parser {,no-}expr-simplify`
When apparmor_parser is used with `-O no-expr-simplify` it can
go into a expoential memory/cpu usage pattern when there are
lines with `/**/`. We dealt with that in the past by adding
special helpers like `AddParametricSnippet`, see 9c74887908b.

But it seems we have reached the limits for this now as we have
various interfaces that e.g. contain `/sys/devices/**/.../**` like
patterns which when combined seems to cause issues that make the
parser easily consume more than 1G of memory.

This commit has an example snap that combines some of the problematic
interfaces and demonstrates that it cannot be compiled with less
than 900M (the real number is probably closer to 1G) when
`no-expr-simplify` is used. It then shows that it does work with
the default apparmor options.