gadget: relax gadget update checks
Don't care anymore if there is a change in offset-write, as that does
not imply that the partitions layout is incompatible. Also, allow
changes in the partition type as long as there is an intersection in
the list of supported types. This will make easier UC20 to UC 22
updates for images using pc gadget.