store: remove refresh hold data in store requests for older proxies
"held" is supported starting v54 of the store API. Some store proxies might
be running older versions of the store which will fail with: "Additional
properties are not allowed ('held' was unexpected)".

This approach retries bad requests, with the new knowledge of the store
version parsed from the response header ""Snap-Store-Version"" and removes
the "held" field if store version is less than v54.

This approach is generic and can be used for other new fields using
minStoreVersion(storeVer, <MIN_VER>).

* o/snapstate: make debug message consistent and store version check explicit

* o/snapstate: check Snap-Store-Version header is valid for retry logic and fix typo in test name

* o/snapstate: extract refresh hold retrieval into a helper function

* o/snapstate: move effectiveRefreshHold to autorefresh.go

* o/snapstate: bump store api requirement to v55 for held field support

	* v55 includes bug fixes related to held field

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>