install: extract unmount helper function
In the review of PR#12939 extracting a common helper for unmounting
was suggested. This commit extracts a new `unmountWithFallbackToLazy`
helper that first tries to do a normal unmount and then falls
back to `syscall.MNT_DETACH` which is the same as `umount --lazy`.
It will ensure the mountpoint will be released in the kernel once
there are no longer any processes keeping it busy.