interfaces: update comments about `no-expr-simplify`
This commit is a followup for the comments in the review for
PR#12918 [1]. As snapd is no longer using `no-expr-simplify`
all reference in the code got updated.

The `AddParametricSnippet()` that was originally introduced to
avoid exponential memory growth with the parser is still useful
and needed because the measurements in PR#12943 show that
multiple lines with overlaping patterns will slow down the
apparmor_parser when running with default `expr-simplify`.

The comments are updated accordingly in this commit.

[1] https://github.com/snapcore/snapd/pull/12918#pullrequestreview-1499831667