osutil: tweak error handling in sys.RunAsUidGid()
When an error happens during the restore of the uid/gid in RunAsUidGid()
deal with that more gracefully. Instead of panic() it is enough to
return the error but *not* call `runtime.UnlockOSThread()`. Because
the uid/gid is per OS-thread on restore failure it is enough to let
the thread "die" which will happen automatically when
runtime.UnlockOSThread is not called.

Also adding a test that ensures that on failure the thread is not
reused and the other threads have the correct uid.