s/seedtest: remove error assertion in makeAssertedSnap when adding snap declarations
if makeAssertedSnap is used to create multiple revisions for the same snap, this
error assertion would trigger since it attempts to add an already existing snap declaration.
Removing it allows makeAssertedSnap to be used for these cases. For example this was
needed for the upgrade scenarios test cases to work in the initramfs verity
mounts suite where we are creating actual asserted snaps.