snap-seccomp: fix tests for explicit/implicit denials
The existing tests for negative filtering did not really test
what we wanted them to test because the seccomp filtering in the
kernel seems to be clever enough to see that e.g. "setpriority"
was not part of the allow list so it never reached the explicit
deny list. This commit fixes this and updates the tests to actually
check for implicit/explicit denials.