wrappers: tell user systemd instances to reload after updating snapd's user units
The snapd.session-agent.service unit written out on Core devices
includes a line like:

    ExecStart=/snap/snapd/NNNN/usr/bin/snap userd --agent

Running user session systemd instances will cache the unit, and should
that snapd revision be removed, the unit can no longer be started. This
can then break future operations that rely on the session agent.

To fix this, we ask all available session agents to do a deamon-reload
after the new versions of the units have been written out. The exec path
in the cached version of the unit should work for this call, since the
old snapd revision is still available while upgrading to the new one.