epoll: make epoll package thread safe
Replace the instance-specific event buffer with a mutex-guarded
registered FD count, allowing an event buffer with a number of slots
equal to the current number of registered FDs to be allocated in
`Wait()`, and thus allowing events from all FDs to be handled by a
single `Wait()` call.

Previously, the instance-specific event buffer was shared by separate
calls to `Wait()`, with the address of the start of the buffer passed
into the EPOLL_WAIT syscall, so multiple concurrent calls to `Wait()`
could cause race conditions where the contents of the buffer was
overwritten by multiple threads at once. Now, since the event buffer is
allocated within the `Wait()` function, this race condition is avoided.

Whenever a new FD is registered, the registeredFdCount variable is
incremented, and whenever an FD is deregistered, that variable is
decremented, in both cases guarded by the mutex for the epoll instance.

So what happens if there are n FDs registered, a call to `Wait()`
begins, and then while waiting, another FD is registered with the epoll
instance? Could this result in a buffer with too few entries for the now
(n+1) registered FDs? No, because beyond race conditions where the
kernel receives two write syscalls simultaneously and these end up in
the same epoll response (extremely unlikely), the only way that multiple
events are returned by the same call to `Wait()` is if there were
previous write events which had been caught by epoll but hadn't yet been
handled with an EPOLL_WAIT syscall. In this case, `Wait()` would
immediately return all those previous events, before the new FD is able
to be registered and written to.

The way around this would be if `Wait()` began, allocated the buffer
with size n, and then the thread were interrupted, and the new FD
registered and written to before the EPOLL_WAIT syscall occurred. In
this case, if all the initial n FDs had pending activity, then the
`Wait()` call would return an event buffer of size n, presumably with
the events for the original n FDs, and a subsequent call to `Wait()`
would return the event on the newly-registered FD. No event is lost in
any case.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>