client: fix specificErr handling in SnapshotExport()
The code was checking generating a `specificErr` but then checked
`if err != nil {` (and not check for specificErr) and also lacked
a unit test.

When adding the code test it became clear that the other issue
there is that the response body was never read so r.err() would
always return nil.

This commit fixes both issues.