epoll: added handling for waiting with no registered FDs
It may be the case that one wishes to wait for epoll events without
first registering any file descriptors to which to listen.  Perhaps
those file descriptors will be registered later by another thread, and
we want to capture any activity on them as soon as that occurs.  In any
case, we do not wish to deny the ability to call `Wait()` if there are
no registered FDs.

However, the `EpollWait` syscall requires a nonzero-length buffer to be
passed in (technically a nonzero length value, along with a pointer to
the buffer). Thus, in order to allow future registered FDs to be handled
by an existing wait call which was initiated when no FDs were
registered, it is necessary to ensure that a buffer of at least length 1
is passed into the `EpollWait` syscall.

Unit tests were added to check for correct waiting behavior when no FDs
are registered, when waiting is initiated before FD is registered, and
when the final FD is deregistered after wait has been initiated.

The `TestWaitThenRegister` function was rewritten to be more salient. In
particular, it now tests behavior when no FDs are registered at the time
`Wait()` is called, and then a new FD is registered and activity occurs
on it.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>