snap: support install/refresh that result in "Wait" changes in the cli better (#12910)
* snap: support install/refresh that result in "Wait" changes in the cli better

On a classic system that supports mode the user experience when refreshing
a snap in wait state (like the kernel) is not great. The progress meter just
displays that a "Re-refresh" is in doing state. This is confusing for the
user as really a reboot is required.

So this commit change the behavior so that after a refresh that
ends up in a "Wait" snap will show a message:
```
Task set to wait until a manual system restart allows to continue
```

* snap: rename waitWaitTasks->waitForTasksInWaitStatus

* move "wait" handling ealier in showDone() and add missing unit test

* snap: remove warning about pressing "ctrl-c" as it no longer applies (thanks to Samuele)

* tests/lib/tools: use correct unit switch

* t/core/snapd-failover: use SIGKILL instead of SIGSEGV

We should avoid triggering of the segmentation fault check

* t/regression/mount-order-regression: set no mem limit

Unfortunately it happens that we hit the memory limit while setting up security profiles for the snap. Set no memory limit for this test

---------

Co-authored-by: Philip Meulengracht <philip.meulengracht@canonical.com>