interfaces/audio: allow to provide pipewire/pulse slots from a containerized pipewire/desktop (#12865)
* UDENG-535: allow access to pipewire socket

This PR ports UDENG-418 to upstream, that enables audio in core
Desktop. It allows access to the pipewire socket from snapped
applications, and allows to have pipewire running inside a snap
container. To avoid security issues, it only allows to access
the sockets located at the folder belonging to the snap that
owns the slot.

With this patch, audio works fine in Core Desktop, and also
screen sharing.

* Replace PLUG_SECURITY_TAGS with SLOT_SECURITY_TAGS

* Remove unneeded "ImplicitOnCore: false" statements

* Better explanation for the /etc/pulse rule

* Add pulseaudio tests