aspects: parse and validate maps and simple strings
Adds support for parsing and validating maps with "schema", "required",
"keys" and "values" constraints. Also adds support for simple strings
(without any of the constraints defined in the spec) because it
simplifies testing of the maps (and, in the case of "keys", it's even
required because they must be string-based).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>