i/prompting: add package for prompting common types/functions
Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

i/prompting: make `TimestampToTime()` return a local time, matching `time.Now()` and `time.Unix()`

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

i/prompting: make duration a string parsable by `time.ParseDuration()`

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

i/prompting: reorder functions

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

i/prompting: remove "app" from prompts and rules

AppArmor permissions are granted by interfaces according to plugs/slots
which may or may not differ between apps within a given snap. For the
home interface, for example, there is never an instance where one app
within a snap is granted a permission that another app in that same snap
is not. Thus, it does not make sense for a rule to apply to only a
single app within a snap.

Instead, the interface field should optionally include the name of the
plug or slot which granted the permissions, which should be received
from the kernel using message tagging. The result should be of the form:

`<interface name>[/<plug or slot>]`

As such, rules should no longer be associated with a particular app, and
in order to match, neither should request prompts. Adjustments are made
to the prompting API to remove the "app" parameter when querying rules.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

i/prompting: add validation functions for expiration

Adds expiration-related analogues for `ValidateLifespanParseDuration`
and `ValidateConstraintsOutcomeLifespanDuration`.

Additionally, improves error messages by including the contents of the
invalid field in the message.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

i/prompting: move prompting common types/functions to interfaces/prompting

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>