sandbox/apparmor/notify: adjusted `Ioctl()` return behavior
In particular, rather than modifying the internal buffer of the
`IoctlRequestBuffer` which was passed into `Ioctl()`, that buffer is now
left at its full size so that it can be reused for future `Ioctl()`
calls. Instead, a new slice with the size returned by the ioctl syscall
and starting at the beginning of the buffer is returned, along with the
an error if the syscall returned an error or an invalid size.  Since the
returned buffer is a slice over the existing `IoctlRequestBuffer`, it
should not be modified, nor should the `IoctlRequestBuffer`'s internal
buffer be modified before any results of the previous call are handled.

Also removed `ReadMessage()`, as it no longer does anything besides
allocate an `IoctlRequestBuffer` and then call `Ioctl()` on it.
Additionally, using `NewIoctlRequestBuffer()` and `Ioctl()` directly
allows for much easier mocking in users of the `notify` package.  Thus,
to avoid confusion, `ReadMessage()` has been removed.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>