cmd/snaplock/runinhibit: add snap revision to inhibit lock file
This is handy to be used by "snap run" to get snap info when
the "current" symlink is not available (i.e. during a refresh
after the snap is unlinked).

Inhibit file (v2) is now JSON instead of the single hint string. This is
needed to carry more information that is needed by "snap run" while
the snap is inhibited. It can be expanded to add more data as needed.

Signed-off-by: Zeyad Gouda <zeyad.gouda@canonical.com>