i/b/shared-memory: error on invalid plug attribute
The isPrivate() method assumes that a plug has been sanitised by
BeforePreparePlug and panics if there's no "private" attribute.
There's been a customer report of this happening and even though the
it's unclear how the attribute is missing/wrong, we should handle it
as gracefully as we can instead of panicking.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>