devicestate: re-init dirs.SetRootDir after mocking ReleaseInfo
When MockReleaseInfo is called the dirs.SetRootDir() must be
called again because when running on e.g. fedora the
dirs.SnapMountDir is set to /var/lib/snapd/snaps but when
mocking a different release dirs.SnapMountDir may not actually
point to /snap leading to test failures.

This commit fixes this by calling dirs.SetRootDir() again
in the firstboot test.