gadget: fix TestMatchDisksToGadgetVolumes on systems with /dev/vda2
The `TestMatchDisksToGadgetVolumes` is using insufficient mocking
and therefor the test will fail on any system that has a /dev/vda2
device (like our ppc64el sbuild environment).

This commit fixes this by splitting the test into three tests that
check different aspects of the `MatchDisksToGadgetVolumes()`.

With that the "not-found" base can be tested easily by ensuring
that a device that is no longer supported by the kernel is used
(/dev/xda which got deleted in v3.9 [0]. There is now a little
bit more duplication but it seems an acceptable trade-off.

[0] https://docs.kernel.org/admin-guide/devices.html