o/d/remodel.go: release lock when resealing
Resealing can take more than a minute since it must hash a lot of files
that it might need to unpack from snaps. It also needs to interact
with the TPM and parse the event log, which might be slow. If
we keep the lock, then snapd will not respond for a long time.

This fixes failures for `tests/nested/manual/core20-remodel`