data/systemd: make snapd depend on snapd.apparmor
On systems with apparmor, there will be issues in snapd if
`snapd.apparmor.service` is not started. For that reason we should
not allow `snapd.apparmor.service` to be disabled.

`snapd.appamor.service` will be skipped automatically if apparmor is
not available with a condition in the unit. So it is safe to always
want it.

See bug https://bugs.launchpad.net/snapd/+bug/1806135