snapfile: show more context when a snap file header cannot be read
There was a recent salesforce case where a invalid snap was
downloaded but because the error from the snapfile only contains
the "bytes" representation it was not clear right away that
the header was actually a json error like: `{"err...`.

This commit tweaks the error so that more context (15 bytes)
are displayed and they are also displayed as quoted string
not only as bytes.