Fix docker-smoke test
A reboot is needed to make sure the networking is properly configured
after removing the docker snap.

Interface docker0 is not removed when docker is uninstalled, this
problem is reproduced with docker snap and deb packages.

Rules added in iptables also are not working properly after docker is
removed. A reboot is recommended.