osutil/epoll: make `WaitTimeout()` reference initial time
Previously, the elapsed time was computed as the total time spent
waiting on the `epoll_wait` syscall, but when waiting for a timeout, one
is generally more concerned with the time which has elapsed since
initiating the call, rather than the total time internally spent waiting
on the syscall (to the extent that can even be accurately measured).

This commit ensures that the function returns once the time since the
start of the call exceeds the given duration.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>