osutil/epoll: make `IsClosed()` use `closed` channel
Thus, it is not necessary to set `e.fd = -1` when the channel is closed,
which is redundant information with the `closed` channel and meant that
reads and writes of `e.fd` needed to be atomic. Now, `e.fd` is never
changed after the epoll instance is created, so there is no need for
atomic reads/writes of `e.fd`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>