osutil/epoll: added `IsClosed()` checks after syscalls
If the epoll instance is closed during a syscall, this could cause the
syscall to return EBADF.  However, since closing the epoll instance is
the cause of this, we want to return `ErrEpollClosed` instead of
`EBADF`.

By checking `IsClosed()` immediately after the syscall returns (and
before checking the error returned by the syscall), we are guaranteed to
return `ErrEpollClosed` if the epoll instance was closed during a
syscall, regardless of whether closing the instance was the cause of an
error during the syscall.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>