osutil/epoll: Only return ErrEpollClosed if error occurred
For methods which invoke syscalls, the epoll instance may be closed
during the syscall, which may cause the syscall to return an error.

Previously, we checked (after the syscall returned) if the epoll
instance was now closed, and if so, returned ErrEpollClosed.  However,
there is a race between the end of the syscall and the check for whether
the epoll instance has been closed, so it may be that the syscall
completed successfully, either because the close occurred after the
syscall completed, or because the close did not trigger an error in the
syscall.  Regardless, we should not return an error if no error actually
occurred.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>